do not reinstall chef everytime

+++ bootstrap-web.sh
@@ -7,7 +7,7 @@ echo '10.0.0.2 chef chef' >> /etc/hosts
 wget http://10.0.0.2:800/darkstar-validator.pem -O /etc/chef/darkstar-validator.pem
 
 # Install chef
-curl -L https://omnitruck.chef.io/install.sh | bash || error_exit 'could not install chef'
+[[ ! -f '/etc/chef/client.rb' ]] && curl -L https://omnitruck.chef.io/install.sh | bash || error_exit 'could not install chef'
 
 # Create first-boot.json
 cat > "/etc/chef/first-boot.json" << EOF
